requirements.txt
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.4.2
Authlib==0.15.3
autopep8==1.5.4
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.8
distlib==0.3.0
distro==1.4.0
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
flask-paginate==0.7.1
Flask-RESTful==0.3.8
Flask-Session==0.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
html5lib==1.0.1
idna==2.10
install==1.3.4
ipaddr==2.2.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2020.4
requests==2.24.0
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.12.1
WTForms==2.3.3
